renamed All to Any
